+++ src/Engine/Globals.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+
+namespace Global
+{
+	const std::string ENGINE_VERSION = "Alpha 0.01";
+};