+++ src/Engine/Timer.h
@@ -1,0 +1,22 @@
+#include "SDL2/SDL.h"
+
+class Timer
+{
+public:
+	Timer();
+
+	// Actions
+	virtual void Start();
+	virtual void Stop();
+	virtual void Pause();
+	virtual void Resume();
+	Uint32 GetTicks();
+
+	bool IsStarted();
+	bool IsActive();
+	bool IsPaused();
+
+private:
+	Uint32 StartTicks, PausedTicks;
+	bool bPaused, bStarted;
+};