+++ src/TestEntity.h
@@ -7,4 +7,7 @@ class TestEntity : public Entity
 public:
 	void Render();
 	void Input();
+	void Tick() override;
+
+	Vector2f velocity;
 };